app/requirements.txt
fastapi
uvicorn==0.29.0
starlette>=0.37.2
celery
redis
python-multipart
pymupdf       # PDF text extraction
pdf2image       # PDF to images
pytesseract     # OCR
pillow        # Image processing
openai       # Markdown formatting
google-generativeai
Mistral
mistralai
Spire.Pdf
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
pymongo>=4.0
motor>=3.1.1
click==8.1.7 
h11==0.14.0